+++ packages/mask/src/styles.tsx
@@ -3,7 +3,7 @@
 export type StyleFn = (
   props: { [key: string]: any },
   state?: { [key: string]: any }
-) => React.CSSProperties
+) => React.CSSProperties & { rx?: number }
 
 // Define the different style functions available for different parts of the Mask component.
 export type Styles = {
@@ -16,7 +16,9 @@ export type Styles = {
 }
 
 // A type for a potentially partial styles object.
-export type StylesObj = Partial<Styles>
+export type StylesObj = {
+  [key in StyleKey]?: StyleFn
+}
 
 // A type for valid keys of the Styles type.
 export type StyleKey = keyof Styles
@@ -27,7 +29,6 @@ export type StyleKey = keyof Styles
  */
 export type StylesKeys = StyleKey
 
-// Define default styles for the Mask component. These styles can be overridden when using the component.
 export const defaultStyles: Styles = {
   maskWrapper: () => ({
     opacity: 0.7,
@@ -38,11 +39,11 @@ export const defaultStyles: Styles = {
     pointerEvents: 'none',
     color: '#000',
   }),
-  svgWrapper: ({ windowWidth, windowHeight, wpx, wpy }) => ({
+  svgWrapper: ({ windowWidth, windowHeight, wpt, wpl }) => ({
     width: windowWidth,
     height: windowHeight,
-    left: Number(wpx) / 2,
-    top: Number(wpy) / 2,
+    left: Number(wpl),
+    top: Number(wpt),
     position: 'fixed',
   }),
   maskArea: ({ x, y, width, height }) => ({
@@ -82,12 +83,15 @@ export const defaultStyles: Styles = {
 }
 
 // Type for a function that gets styles for a given key and extra data.
-export type getStylesType = (key: StylesKeys, extra?: any) => {}
+export type getStylesType = (
+  key: StyleKey,
+  extra?: { [key: string]: any }
+) => {}
 
 // A utility function to determine the appropriate styles based on the given key and state.
 // If a custom style function exists for the key, it's used. Otherwise, default styles are returned.
 export function stylesMatcher(styles: StylesObj) {
-  return (key: StyleKey, state: {}): any => {
+  return (key: StyleKey, state: {}): React.CSSProperties & { rx?: number } => {
     const base = defaultStyles[key](state)
     const custom = styles[key]
     return custom ? custom(base, state) : base