+++ packages/popover/src/index.tsx
@@ -1,7 +1,6 @@
 import Popover from './Popover'
 import { PositionType, PositionProps } from './Popover'
 import { StylesObj } from './styles'
-import './styles.css'
 
 export default Popover
 export type { PositionType as Position }