+++ packages/mask/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @react-explore-kit/mask
 
+## 7.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/utils@7.0.0
+
+## 6.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/utils@6.0.0
+
+## 5.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/utils@5.0.0
+
+## 4.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/utils@4.0.0
+
+## 3.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/utils@3.0.0
+
 ## 2.0.0
 
 ### Major Changes