+++ packages/tour/CHANGELOG.md
@@ -1,5 +1,70 @@
 # @react-explore-kit/tour
 
+## 7.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/popover@7.0.0
+  - @react-explore-kit/utils@7.0.0
+  - @react-explore-kit/mask@7.0.0
+
+## 6.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/popover@6.0.0
+  - @react-explore-kit/utils@6.0.0
+  - @react-explore-kit/mask@6.0.0
+
+## 5.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/popover@5.0.0
+  - @react-explore-kit/utils@5.0.0
+  - @react-explore-kit/mask@5.0.0
+
+## 4.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/popover@4.0.0
+  - @react-explore-kit/utils@4.0.0
+  - @react-explore-kit/mask@4.0.0
+
+## 3.0.0
+
+### Major Changes
+
+- v2
+
+### Patch Changes
+
+- Updated dependencies
+  - @react-explore-kit/popover@3.0.0
+  - @react-explore-kit/utils@3.0.0
+  - @react-explore-kit/mask@3.0.0
+
 ## 2.0.0
 
 ### Major Changes