+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @react-explore-kit/utils
 
+## 7.0.0
+
+### Major Changes
+
+- v2
+
+## 6.0.0
+
+### Major Changes
+
+- v2
+
+## 5.0.0
+
+### Major Changes
+
+- v2
+
+## 4.0.0
+
+### Major Changes
+
+- v2
+
+## 3.0.0
+
+### Major Changes
+
+- v2
+
 ## 2.0.0
 
 ### Major Changes